requirements.txt
feedparser
pandas
tqdm
python-dateutil
pyarrow
